Refactor ProductsPage: Simplify state management by removing filters and cart handling, and enhance category display with links
